lesson6/notifier/requirements.txt
pyTelegramBotAPI==4.23.0
python-dotenv
loguru
dependency_injector
pika